+++ components/legislator/LegislatorComponents.tsx
@@ -54,6 +54,24 @@ export function PartyLabel(props: { party: string }) {
   }
 }
 
+export const formatPhoneNumber = (value: string) => {
+  if (!value) return value
+
+  const phoneNumber = value.replace(/[^\d]/g, "")
+  const phoneNumberLength = phoneNumber.length
+
+  // Format as (XXX) XXX-XXXX
+  if (phoneNumberLength < 4) return phoneNumber
+  if (phoneNumberLength < 7) {
+    return `(${phoneNumber.slice(0, 3)}) ${phoneNumber.slice(3)}`
+  }
+  return `
+      (${phoneNumber.slice(0, 3)})
+       ${phoneNumber.slice(3, 6)}-
+       ${phoneNumber.slice(6, 10)}
+    `
+}
+
 /** Social Media components **/
 
 export function Bluesky() {
